Added webpage rendering with rich
